Application Program for Subject-Oriented Thinking and Programming in My University 1st Grade
It was a semester project for the second semester of the first grade. We have created a program that applies to students applying for the classes they want so that they can pick the classes they want every semester. It's my first Java program, so I'm not very good at it

We have created a program that applies to students applying for the classes they want so that they can pick the classes they want every semester. It's my first Java program, so I'm not very good at it

#java #programming #university student #1st grade # object oriented thought and programming